Hexen|SndInfoParser: Fix patching of MAPINFO (lump) definitions with SNDINFO-read values
(The only time a mapinfo_t is modified outside of MapInfoParser).

+++ doomsday/plugins/hexen/include/p_mapinfo.h
@@ -30,10 +30,10 @@
 
 typedef struct mapinfo_s {
     uint         map; ///< Logical map number.
-    short        cluster;
+    int          cluster;
     uint         warpTrans;
     uint         nextMap;
-    short        cdTrack;
+    int          cdTrack;
     char         title[32];
     materialid_t sky1Material;
     materialid_t sky2Material;
